requirements.txt
Django>=5.2
djangorestframework
djangorestframework-simplejwt
pandas
numpy
scikit-learn
joblib
bcrypt
psycopg2-binary
